Refresh token

+++ src/pages/UsersPage/useUsersPage.ts
@@ -5,15 +5,20 @@ import { addUser, deleteUser, editUser, getUsers } from './UsersPage.api';
 import { openNotification } from 'src/utils/utils';
 import { useCallback } from 'react';
 import { IUserCreate, IUserDTO } from 'src/user/user.types';
+import { useAuth } from 'src/hooks/useAuth';
+import { postRefreshToken } from '../LoginPage/LoginPage.api';
+import { useNavigate } from 'react-router';
 
 export const useUsersPage = () => {
+  const { auth } = useAuth();
+  const navigate = useNavigate();
   const { isLoading, data, refetch } = useQuery(
     ['getUsers'],
     async () => {
       try {
         return await getUsers();
       } catch (err) {
-        openNotification('error', err.response?.status, err.response?.data);
+        openNotification('error', err.response?.statusCode, err.response?.data);
       }
     },
     { ...reactQueryNoCacheOptions },
@@ -24,25 +29,60 @@ export const useUsersPage = () => {
       await addUser(user);
       await refetch();
     } catch (err) {
-      openNotification('error', err.response?.status, err.response?.data);
+      openNotification('error', err.response?.statusCode, err.response?.data);
     }
   }, []);
 
   const handleEditUser = useCallback(async (data: IUserDTO) => {
     try {
-      await editUser(data, { withCredentials: true });
+      const response = await editUser(data, { withCredentials: true });
+
+      if (response.data?.statusCode == 403) {
+        const refreshResponse = await postRefreshToken(
+          { refreshToken: auth.refreshToken },
+          { withCredentials: true },
+        );
+
+        if (refreshResponse.data?.statusCode === 401) {
+          navigate('/login');
+          return;
+        }
+
+        await editUser(data, { withCredentials: true });
+        await refetch();
+
+        return;
+      }
+
       await refetch();
     } catch (err) {
-      openNotification('error', err.response?.status, err.response?.data);
+      openNotification('error', err.response?.statusCode, err.response?.data);
     }
   }, []);
 
   const handleDeleteUser = useCallback(async (id: string) => {
     try {
-      await deleteUser(id, { withCredentials: true });
+      const response = await deleteUser(id, { withCredentials: true });
+      console.log(response);
+      if (response.data?.statusCode == 403) {
+        const refreshResponse = await postRefreshToken(
+          { refreshToken: auth.refreshToken },
+          { withCredentials: true },
+        );
+
+        if (refreshResponse.data?.statusCode === 401) {
+          navigate('/login');
+          return;
+        }
+
+        await deleteUser(id, { withCredentials: true });
+        await refetch();
+
+        return;
+      }
       await refetch();
     } catch (err) {
-      openNotification('error', err.response?.status, err.response?.data);
+      openNotification('error', err.response?.statusCode, err.response?.data);
     }
   }, []);
